app/requirements.txt
sentence-transformers
opensearch-py
pandas
python-dateutil
fastapi
uvicorn[standard]

